refactor: move processBillingReceipt to PaymentContainer
This removes the usage of multiple window and document usages
